+++ programs/mango-v4/src/instructions/token_liq_bankruptcy.rs
@@ -9,8 +9,7 @@ use crate::state::*;
 
 use crate::accounts_ix::*;
 use crate::logs::{
-    LoanOriginationFeeInstruction, TokenBalanceLog, TokenLiqBankruptcyLog,
-    WithdrawLoanOriginationFeeLog,
+    LoanOriginationFeeInstruction, TokenBalanceLog, TokenLiqBankruptcyLog, WithdrawLoanLog,
 };
 
 pub fn token_liq_bankruptcy(
@@ -149,7 +148,7 @@ pub fn token_liq_bankruptcy(
             // transfer liab from liqee to liqor
             let (liqor_liab, liqor_liab_raw_token_index, _) =
                 liqor.ensure_token_position(liab_token_index)?;
-            let (liqor_liab_active, loan_origination_fee) =
+            let liqor_liab_withdraw_result =
                 liab_bank.withdraw_with_fee(liqor_liab, liab_transfer, now_ts)?;
 
             // liqor liab
@@ -169,13 +168,18 @@ pub fn token_liq_bankruptcy(
                 require!(liqor_health >= 0, MangoError::HealthMustBePositive);
             }
 
-            if loan_origination_fee.is_positive() {
-                emit!(WithdrawLoanOriginationFeeLog {
+            if liqor_liab_withdraw_result
+                .loan_origination_fee
+                .is_positive()
+            {
+                emit!(WithdrawLoanLog {
                     mango_group: ctx.accounts.group.key(),
                     mango_account: ctx.accounts.liqor.key(),
                     token_index: liab_token_index,
-                    loan_origination_fee: loan_origination_fee.to_bits(),
-                    instruction: LoanOriginationFeeInstruction::LiqTokenBankruptcy
+                    loan_amount: liqor_liab_withdraw_result.loan_amount.to_bits(),
+                    loan_origination_fee: liqor_liab_withdraw_result.loan_origination_fee.to_bits(),
+                    instruction: LoanOriginationFeeInstruction::LiqTokenBankruptcy,
+                    price: Some(liab_oracle_price.to_bits())
                 });
             }
 
@@ -185,7 +189,7 @@ pub fn token_liq_bankruptcy(
                     ctx.accounts.liqor.key(),
                 );
             }
-            if !liqor_liab_active {
+            if !liqor_liab_withdraw_result.position_is_active {
                 liqor.deactivate_token_position_and_log(
                     liqor_liab_raw_token_index,
                     ctx.accounts.liqor.key(),