+++ programs/mango-v4/src/logs.rs
@@ -259,6 +259,17 @@ pub struct WithdrawLoanOriginationFeeLog {
 }
 
 #[event]
+pub struct WithdrawLoanLog {
+    pub mango_group: Pubkey,
+    pub mango_account: Pubkey,
+    pub token_index: u16,
+    pub loan_amount: i128,
+    pub loan_origination_fee: i128,
+    pub instruction: LoanOriginationFeeInstruction,
+    pub price: Option<i128>, // Ideally would log price everywhere but in serum3_settle_funds oracle is not a passed in account
+}
+
+#[event]
 pub struct TokenLiqBankruptcyLog {
     pub mango_group: Pubkey,
     pub liqee: Pubkey,