+++ programs/mango-v4/src/state/mango_account.rs
@@ -14,7 +14,6 @@ use crate::error::*;
 use crate::health::{HealthCache, HealthType};
 use crate::logs::{DeactivatePerpPositionLog, DeactivateTokenPositionLog};
 
-use super::dynamic_account::*;
 use super::BookSideOrderTree;
 use super::FillEvent;
 use super::LeafNode;
@@ -25,6 +24,7 @@ use super::Serum3MarketIndex;
 use super::TokenConditionalSwap;
 use super::TokenIndex;
 use super::FREE_ORDER_SLOT;
+use super::{dynamic_account::*, Group};
 use super::{PerpPosition, Serum3Orders, TokenPosition};
 use super::{Side, SideAndOrderTree};
 
@@ -338,17 +338,22 @@ impl MangoAccountFixed {
     }
 
     /// Add new fees that are usable with the buyback fees feature.
+    ///
+    /// Any call to this should be preceeded by a call to expire_buyback_fees earlier
+    /// in the same instruction.
     pub fn accrue_buyback_fees(&mut self, amount: u64) {
         self.buyback_fees_accrued_current =
             self.buyback_fees_accrued_current.saturating_add(amount);
     }
 
     /// Reduce the available buyback fees amount because it was used up.
+    ///
+    /// Panics if `amount` exceeds the available accrued amount
     pub fn reduce_buyback_fees_accrued(&mut self, amount: u64) {
         if amount > self.buyback_fees_accrued_previous {
-            self.buyback_fees_accrued_current = self
-                .buyback_fees_accrued_current
-                .saturating_sub(amount - self.buyback_fees_accrued_previous);
+            let remaining_amount = amount - self.buyback_fees_accrued_previous;
+            assert!(remaining_amount <= self.buyback_fees_accrued_current);
+            self.buyback_fees_accrued_current -= remaining_amount;
             self.buyback_fees_accrued_previous = 0;
         } else {
             self.buyback_fees_accrued_previous -= amount;
@@ -1076,14 +1081,17 @@ impl<
         perp_market_index: PerpMarketIndex,
         perp_market: &mut PerpMarket,
         fill: &FillEvent,
+        group: &Group,
     ) -> Result<()> {
         let side = fill.taker_side().invert_side();
         let (base_change, quote_change) = fill.base_quote_change(side);
         let quote = I80F48::from(perp_market.quote_lot_size) * I80F48::from(quote_change);
         let fees = quote.abs() * I80F48::from_num(fill.maker_fee);
         if fees.is_positive() {
-            self.fixed_mut()
-                .accrue_buyback_fees(fees.floor().to_num::<u64>());
+            let f = self.fixed_mut();
+            let now_ts = Clock::get().unwrap().unix_timestamp.try_into().unwrap();
+            f.expire_buyback_fees(now_ts, group.buyback_fees_expiry_interval);
+            f.accrue_buyback_fees(fees.floor().to_num::<u64>());
         }
         let pa = self.perp_position_mut(perp_market_index)?;
         pa.settle_funding(perp_market);
@@ -1173,8 +1181,25 @@ impl<
     ) -> Result<()> {
         let post_init_health = health_cache.health(HealthType::Init);
         msg!("post_init_health: {}", post_init_health);
+
+        // Accounts that have negative init health may only take actions that don't further
+        // decrease their health.
+        // To avoid issues with rounding, we allow accounts to decrease their health by up to
+        // $1e-6. This is safe because the grace amount is way less than the cost of a transaction.
+        // And worst case, users can only use this to gradually drive their own account into
+        // liquidation.
+        // There is an exception for accounts with health between $0 and -$0.001 (-1000 native),
+        // because we don't want to allow empty accounts or accounts with extremely tiny deposits
+        // to immediately drive themselves into bankruptcy. (accounts with large deposits can also
+        // be in this health range, but it's really unlikely)
+        let health_does_not_decrease = if post_init_health < -1000 {
+            post_init_health.ceil() >= pre_init_health.ceil()
+        } else {
+            post_init_health >= pre_init_health
+        };
+
         require!(
-            post_init_health >= 0 || post_init_health > pre_init_health,
+            post_init_health >= 0 || health_does_not_decrease,
             MangoError::HealthMustBePositiveOrIncrease
         );
         Ok(())
@@ -1736,7 +1761,7 @@ mod tests {
         assert_eq!(fixed.buyback_fees_expiry_timestamp, 1070);
         assert_eq!(fixed.buyback_fees_accrued(), 12);
 
-        fixed.reduce_buyback_fees_accrued(100);
+        fixed.reduce_buyback_fees_accrued(12);
         assert_eq!(fixed.buyback_fees_accrued(), 0);
     }
 }